💊 Release 1.0.1 - Hotfix
Release 1.0.0 came with a serious bug. By calling fgets() instead of
fread(), the program wouldn't always output a complete password, as
part of the array would be full of zeros.

Additionally, the output would output a password 1 character longer
than requested.

The build systems' changelogs, meson.build, and the README now
reflect the updates and the new version.